+++ client/src/research/InteractionsView.story.js
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 import InteractionsView from './InteractionsView.js';
 
 
-storiesOf('InteractionsView', module) //eslint-disable-line no-undef
+storiesOf('Research/InteractionsView', module) //eslint-disable-line no-undef
   .add('normal', () => {
     const db = require('../../../tmp/swipe-right-db.json');
     const {interactions} = db;
