+++ client/src/research/LoginPage.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+
+
+// The page for users to login for accessing research data.
+class LoginPage extends Component {
+  render() {
+    return <div>login please!</div>;
+  }
+}
+
+export default LoginPage;
