+++ client/src/research/LoginPage.test.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import LoginPage from './LoginPage';
+
+
+it('renders without crashing', async () => {
+  const div = document.createElement('div');
+  await ReactDOM.render(<LoginPage />, div);
+});
