+++ client/src/research/ResearchPage.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import Interactions from './Interactions';
+
+
+// The page for logged-in user to access research data.
+class ResearchPage extends Component {
+  render() {
+    return <Interactions />;
+  }
+}
+
+export default ResearchPage;
